Fix compilation with clang on x86
- disable availability of EBX: fixes compilation error with inline
  assembly in libavcodec/x86/cabac.h
- fixes Wincompatible-pointer-types and Wformat errors in vaapi

+++ libavutil/hwcontext_vaapi.c
@@ -1099,7 +1099,7 @@ static int vaapi_map_from_drm(AVHWFramesContext *src_fc, AVFrame *dst,
     int err, i, j;
 
 #if !VA_CHECK_VERSION(1, 1, 0)
-    unsigned long buffer_handle;
+    uintptr_t buffer_handle;
     VASurfaceAttribExternalBuffers buffer_desc;
     VASurfaceAttrib attrs[2] = {
         {
